finish add recyclerVIew wih its adapter (move LongClick code from MainActivity to Adapter)

+++ app/src/main/java/com/example/jbt/omdb/MainActivity.java
@@ -2,28 +2,22 @@ package com.example.jbt.omdb;
 
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.database.Cursor;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.SparseBooleanArray;
-import android.view.ActionMode;
+import android.support.v7.widget.LinearLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.widget.AbsListView;
-import android.widget.AdapterView;
-import android.widget.ListView;
 import android.widget.Toast;
 
 public class MainActivity extends AppCompatActivity {
 
-    private MovieAdapter mAdapter;
+    private MovieRecyclerAdapter mAdapter;
     private MoviesDBHelper mDbHelper;
-
-    private ListView mListView;
 
     public static final String LOG_CAT = "OMDB:";
 
@@ -32,13 +26,18 @@ public class MainActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        mListView = (ListView)findViewById(R.id.mainListView);
+        RecyclerView recyclerView = (RecyclerView) findViewById(R.id.mainListView);
+
+        if (recyclerView == null )
+            return;
 
         mDbHelper = new MoviesDBHelper(this);
         mDbHelper.deleteAllSearchResult();
 
-        mAdapter = new MovieAdapter(this, mDbHelper.GetDetailsMovieCursor());
-        mListView.setAdapter(mAdapter);
+        mAdapter = new MovieRecyclerAdapter(this, mDbHelper.getDetailsMovieArrayList());
+
+        recyclerView.setAdapter(mAdapter);
+        recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
         FloatingActionButton addFab = (FloatingActionButton) findViewById(R.id.addFAB);
         if (addFab != null)
@@ -74,83 +73,7 @@ public class MainActivity extends AppCompatActivity {
                 }
             });
 
-        mListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
 
-                Cursor c = (Cursor) parent.getItemAtPosition(position);
-                if (c != null) {
-                    Movie movie = Utility.getMovieFromCursor(c);
-                    launchEditActivity(movie);
-                }
-            }
-        });
-
-
-        mListView.setChoiceMode(ListView.CHOICE_MODE_MULTIPLE_MODAL);
-        mListView.setMultiChoiceModeListener(new AbsListView.MultiChoiceModeListener() {
-
-            @Override
-            public void onItemCheckedStateChanged(ActionMode mode, int position, long id, boolean checked) {
-            }
-
-            @Override
-            public boolean onCreateActionMode(ActionMode mode, Menu menu) {
-                MenuInflater inflater = mode.getMenuInflater();
-                inflater.inflate(R.menu.main_context_menu, menu);
-                return true;
-            }
-
-            @Override
-            public boolean onPrepareActionMode(ActionMode mode, Menu menu) {
-                return false;
-            }
-
-            @Override
-            public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
-
-                SparseBooleanArray checked = mListView.getCheckedItemPositions();
-
-                if (!checked.valueAt(0))
-                    return false;
-
-                int position = checked.keyAt(0);
-
-                Cursor c = mAdapter.getCursor();
-
-                if (c == null)
-                    return false;
-
-                c.moveToPosition(position);
-
-                Movie movie = Utility.getMovieFromCursor(c);
-
-                switch (item.getItemId())
-                {
-                    case R.id.editMenuItem:
-                        launchEditActivity(movie);
-                        mode.finish();
-                        return true;
-
-                    case R.id.deleteMenuItem:
-                        if ( mDbHelper.deleteMovie(movie.getId())) {
-                            String movieDeletedMsg = getString(R.string.movie_deleted_msg);
-                            Toast.makeText(MainActivity.this, movieDeletedMsg, Toast.LENGTH_SHORT).show();
-                            refreshMainList();
-                            mode.finish();
-                        }
-                        return true;
-
-                    default:
-                        return false;
-                }
-            }
-
-            @Override
-            public void onDestroyActionMode(ActionMode mode) {
-
-            }
-        });
     }
 
     @Override
@@ -226,16 +149,7 @@ public class MainActivity extends AppCompatActivity {
                 .show();
     }
 
-
-    private void launchEditActivity(Movie movie)
-    {
-        Intent intent = new Intent(MainActivity.this, EditActivity.class);
-        intent.putExtra(WebSearchActivity.INTENT_MOVIE_KEY, movie);
-        startActivity(intent);
-    }
-
-
     private void refreshMainList() {
-        mAdapter.changeCursor(mDbHelper.GetDetailsMovieCursor());
+        mAdapter.setData(mDbHelper.getDetailsMovieArrayList());
     }
 }
